rework examples
- restructure example part of documentation
- adjust existing example files
- add examples files (empty)
