+++ docs/source/pages/examples/CAN/can.rst
@@ -1,0 +1,8 @@
+CAN
+===
+
+.. toctree::
+
+    python-can.rst
+
+.. seealso:: https://github.com/mdabrowski1990/uds/tree/main/examples/can
