+++ examples/can/python-can/kvaser/receive_packet.py
@@ -1,17 +1,17 @@
-import asyncio
 from pprint import pprint
+from threading import Timer
 
 from can import Bus, Message
 from uds.can import CanAddressingFormat, CanAddressingInformation
 from uds.transport_interface import PyCanTransportInterface
 
 
-async def main():
+def main():
     # configure CAN interfaces
-    kvaser_interface_1 = Bus(interface="kvaser", channel=0, fd=True, receive_own_messages=True)
-    kvaser_interface_2 = Bus(interface="kvaser", channel=1, fd=True, receive_own_messages=True)
+    kvaser_interface_1 = Bus(interface="kvaser", channel=0, fd=True, receive_own_messages=True)  # receiving
+    kvaser_interface_2 = Bus(interface="kvaser", channel=1, fd=True, receive_own_messages=True)  # sending
 
-    # configure Addressing Information of a CAN Node
+    # configure Addressing Information of a CAN Node (example values)
     addressing_information = CanAddressingInformation(
         addressing_format=CanAddressingFormat.NORMAL_11BIT_ADDRESSING,
         tx_physical={"can_id": 0x611},
@@ -29,22 +29,21 @@ async def main():
     frame_3 = Message(arbitration_id=0x612, data=[0x02, 0x3E, 0x00, 0xAA, 0xAA, 0xAA, 0xAA, 0xAA])
 
     # receive CAN packet 1
-    kvaser_interface_2.send(frame_1)  # transmit CAN Frame 1
-    record_1 = await can_ti.async_receive_packet(timeout=1000)   # receive CAN packet 1 carried by frame 1
+    Timer(interval=0.1, function=kvaser_interface_1.send, args=(frame_1,)).start()  # schedule transmission of frame 1
+    record_1 = can_ti.receive_packet(timeout=1000)  # receive CAN packet 1 carried by frame 1
     pprint(record_1.__dict__)  # show attributes of CAN packet record 1
 
     # receive CAN packet 2
-    kvaser_interface_2.send(frame_2)  # transmit CAN Frame 2
-    kvaser_interface_2.send(frame_3)  # transmit CAN Frame 3
-    record_2 = await can_ti.async_receive_packet(timeout=1000)
+    Timer(interval=0.1, function=kvaser_interface_1.send, args=(frame_2,)).start()  # schedule transmission of frame 2
+    Timer(interval=0.2, function=kvaser_interface_1.send, args=(frame_3,)).start()  # schedule transmission of frame 3
+    record_2 = can_ti.receive_packet(timeout=1000)  # receive CAN packet 2 carried by frame 3
     pprint(record_2.__dict__)  # show attributes of CAN packet record 2
 
     # close connections with CAN interfaces
     del can_ti
-    await asyncio.sleep(0.1)  # wait to make sure all tasks are closed
     kvaser_interface_1.shutdown()
     kvaser_interface_2.shutdown()
 
 
 if __name__ == "__main__":
-    asyncio.run(main())
+    main()
